+++ CHANGES.rst
@@ -8,6 +8,7 @@ in progress
 - Basic implementation generating UUIDv4 identifiers
 - Add ULID identifiers
 - Add ``--upper`` and ``--lower`` options for output transformation
+- Add Nagamani19 identifiers
 
 
 2019-06-28 0.0.0